+++ WikiCrawler/Program.cs
@@ -10,8 +10,8 @@ namespace WikiCrawler
 			//string croppath = "uwash_images/53_cropped.jpg";
 			//ImageUtils.AutoCropJpg("uwash_images/53.jpg", croppath, 0xffffffff, 0.97f, 5);
 
-			//try
-			//{
+			try
+			{
 				Console.Write("Task>");
 				string task = Console.ReadLine();
 
@@ -76,8 +76,8 @@ namespace WikiCrawler
 				//UWashCats.Do();
 				//SingleUpload.Do();
 				//NsrwFollowup.Do();
-			/*}
-			catch (Exception e)
+			}
+			/*catch (Exception e)
 			{
 				Console.WriteLine(e);
 				using (StreamWriter writer = new StreamWriter(new FileStream("errorlog.txt", FileMode.Create)))
@@ -85,6 +85,11 @@ namespace WikiCrawler
 					writer.WriteLine(e.ToString());
 				}
 			}*/
+			finally
+			{
+				//TODO: check dirty
+				CategoryTranslation.SaveOut();
+			}
 
 			Console.WriteLine("Done");
 			Console.ReadLine();
