+++ WikiCrawler/UWash/DataStructures/UWashProjectConfig.cs
@@ -6,25 +6,13 @@ using System.Threading.Tasks;
 
 namespace UWash
 {
-	public class UWashProjectConfig
+	public class UWashProjectConfig : ProjectConfig
 	{
-		public string informationTemplate = "Photograph";
-		public string defaultAuthor;
-		public string defaultPubCountry;
-		public string masterCategory;
-		public string checkCategory;
-		public string[] additionalCategories;
-		public string filenameSuffix;
 		public string digitalCollectionsKey;
 		public string sourceTemplate;
 		public int minIndex;
 		public int maxIndex;
 
 		public bool allowCrop = true;
-		public bool allowFailedCreators = false;
-
-		public bool allowUpload = true;
-		public bool allowDataDownload = true;
-		public bool allowImageDownload = true;
 	}
 }
