+++ WikiCrawler/WikiCrawler.csproj
@@ -37,16 +37,13 @@
     <Prefer32Bit>false</Prefer32Bit>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="FreeImageNET">
-      <HintPath>E:\Manual Program Files\FreeImage\Wrapper\FreeImage.NET\cs\Library\bin\Release\FreeImageNET.dll</HintPath>
-    </Reference>
     <Reference Include="MySql.Data, Version=6.9.9.0, Culture=neutral, PublicKeyToken=c5687fc88969c44d, processorArchitecture=MSIL" />
     <Reference Include="MySql.Data.Entity.EF5, Version=6.9.9.0, Culture=neutral, PublicKeyToken=c5687fc88969c44d, processorArchitecture=MSIL" />
     <Reference Include="MySql.Data.Entity.EF6, Version=6.9.9.0, Culture=neutral, PublicKeyToken=c5687fc88969c44d, processorArchitecture=MSIL" />
     <Reference Include="MySql.Fabric.Plugin, Version=6.9.9.0, Culture=neutral, PublicKeyToken=c5687fc88969c44d, processorArchitecture=MSIL" />
     <Reference Include="MySql.Web, Version=6.9.9.0, Culture=neutral, PublicKeyToken=c5687fc88969c44d, processorArchitecture=MSIL" />
     <Reference Include="Newtonsoft.Json">
-      <HintPath>E:\Manual Program Files\JSON.NET\Bin\net40\Newtonsoft.Json.dll</HintPath>
+      <HintPath>..\..\..\Manual Program Files\JSON.NET\Bin\net45\Newtonsoft.Json.dll</HintPath>
     </Reference>
     <Reference Include="System" />
     <Reference Include="System.configuration" />
@@ -61,12 +58,21 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Backend\ImageUtils.cs" />
+    <Compile Include="Batch\BatchDownloader.cs" />
+    <Compile Include="Batch\BatchTask.cs" />
+    <Compile Include="Batch\BatchUploader.cs" />
+    <Compile Include="Batch\ContentDmDownloader.cs" />
+    <Compile Include="Batch\ProjectConfig.cs" />
+    <Compile Include="Batch\UWashDownloader.cs" />
+    <Compile Include="Batch\UWashUploader.cs" />
     <Compile Include="Configuration.cs" />
     <Compile Include="Credentials.cs" />
     <Compile Include="FindCategoryCreators.cs" />
     <Compile Include="LinksNullEditor.cs" />
     <Compile Include="MassDownloader.cs" />
+    <Compile Include="Utility\CreatorUtility.cs" />
     <Compile Include="Utility\Dimensions.cs" />
+    <Compile Include="Utility\LicenseUtility.cs" />
     <Compile Include="Utility\StringUtility.cs" />
     <Compile Include="CategoryTranslation.cs" />
     <Compile Include="CommonsCreatorFromWikidata.cs" />
@@ -76,10 +82,9 @@
     <Compile Include="Tropenmuseum.cs" />
     <Compile Include="Utility\Wayback.cs" />
     <Compile Include="UWash\DataStructures\UWashConfig.cs" />
-    <Compile Include="UWash\DataStructures\UWashCreator.cs" />
+    <Compile Include="Utility\Creator.cs" />
     <Compile Include="UWash\DataStructures\UWashFailure.cs" />
     <Compile Include="UWash\DataStructures\UWashProjectConfig.cs" />
-    <Compile Include="UWash\UWashCountry.cs" />
     <Compile Include="Wikimedia\Date.cs" />
     <Compile Include="Wikimedia\DotNetWikiBot.cs" />
     <Compile Include="Old\Dsea.cs" />
